polls/requirements.txt
Django
djangorestframework
djongo
psycopg2-binary
gunicorn
requests
sqlparse==0.2.4